requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.4
defusedxml==0.7.0rc2
Django==3.1.4
django-3-jet==1.0.8
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-js-asset==1.2.2
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
stripe==2.56.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
